backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-extensions==3.2.3
django-rql==4.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
lark-parser==0.11.0
lib-rql==2.0.2
packaging==24.2
pillow==11.0.0
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
